matching: remove the [check_partial] heuristics
The [check_partial] heuristics are a coarse-grained approach degrade
matching totality information in certain cases -- when clauses contain
both reads of mutable fields and either guards or forcing of lazy
patterns.

It is not quite correct (it is not enough to prevent #7241), and is
not sufficient conceptually for OCaml Multicore. In a Multicore world,
other domains may race to update values concurrently, so we cannot
assume that only a fixed set of matching constructions may result in
side-effects.

This heuristic is subsumed by the recent changes to:
  - make context information accurate by binding mutable fields eagerly
  - make totality information accurate by tracking the mutability of
    the current position
and can now be retired.

Note: the [check_partial] function also contained hidden logic to deal
with the pattern-matching programs with no cases at all (more
precisely, those that have only refutation cases, like `function _ ->
.`), we retain this logic and move it to `toplevel_handler`.