Fix stack handling in POWER SIGTRAP handler (#12482)
The signal handler updates the stack prior to raising the array_bound_error
exception, but assumes this information is available at the bottom of the
stack, regardless of actual stack allocation for local variables.

Fix this by passing the appropriate offset in a temporary register prior
to issueing conditional trap instructions. The signal handler can then
recompute the proper addresses.