ocamldep: use Load_path
The referenced `OCAMLCOMP` environment variable was probably a typo,
referring to `OCAMLCOMPPARAM`, later renamed to `OCAMLPARAM`. The load
path is still reinitialised after the call to [Compenv.readenv], and
the caching is taken care of by `Load_path`.

Entries returned by Load_path need to be sorted against the include
path(s) to ensure modules are not shadowed.