Dynlink: stop configuring warnings in source files
Most of the settings in these files were redundent with those of the build
system anyway. The only warning which was disabled in the source
files but not in the build system was warning 30 (duplicate-definitions)
but since the current implementations do not trigger it anyway it
seems better to not disable it, at least as long as there is no
need to do so.