runtime: caml_domain_is_multicore(void) => caml_runtime_is_multicore(void)
We are trying to enforce the convention that [caml_domain_is_foo]
functions take a domain_state parameter. [caml_domain_is_multicore] is
inherently a function that:
- does not require ownership of a domain state
- was implemented in a way relying on dom_internal rather than domain_state

We rename the function to clarify that it is not a property of the
current domain, change its implementation to ensure that it can be
called in any context, and document its purpose.