Building the compiler with ThreadSanitizer and running the testsuite caused too many reports in OCaml 5 and was disabled (see #11040).
Since then, the work on TSan support for OCaml programs has led to fix a number
of those data races and temporarily silence the ones that are waiting to be
investigated (see #11040 again). As a result, running the testsuite with
`--enable-tsan` is now a cheap and effective way of detecting new data races
that may be introduced in the runtime.

A second good reason to restore the TSan CI is that it will detect early if a
recent change has accidentally broken TSan instrumentation (as has happened
before as an accidental consequence of removing a symbol
https://github.com/ocaml/ocaml/pull/12383#pullrequestreview-1593390354), or
other issues (e.g. a new test revealed a TSan limitation with signals
https://github.com/ocaml/ocaml/pull/12561#issuecomment-1729862795).

Adding this test to the Github Actions CI arguably lengthens the runs (a GHA
run on amd64 Linux with TSan takes about 50 minutes). This PR therefore
suggests the compromise of enabling it on the Inria CI which is run on every
merge.