Fix signed integer overflows when computing min int
The C standard, paragraph 6.5.7-4, reads

> E1 << E2
> If E1 has a signed type and nonnegative value, and E1 × 2^E2 is
> representable in the result type, then that is the resulting value;
> otherwise, the behavior is undefined.

Thus, for a signed type t, 1 << (sizeof(t)-1) is undefined as it
equals to the maximum value representable in type t, plus one.  It
usually overflows and amounts to the minimum value representable in
type t, but that's still UB. Use constants from limits.h, because the
minimum value cannot be easily computed.

Introduce the macro INTNAT_MIN.