Make the build_libraries_manpages configuration variable more accurate
Before this commit, the build_libraries_manpages variable (aka STDLIB_MANPAGES)
was set to true if the build of the manpages for libraries was
enabled at configure time, whether ocamldoc (which is required
 to build those manpages) was enabled or not. It was thus the build
system's responsibility to determine whether the manpages for
libraries should be built / installed,
by testing both the
build_ocamldoc and the build_libraries_manpages variables.

However, it is known at configure time whether ocamldoc is available
or not, which makes it possible to set build_libraries_manpages to
true only if the manpages for libraries have been requested AND
ocamldoc has been enabled.

This is what is done in this commit, leading to a simplification
on the build system's side since it becomes enough to test only
one variable, namely build_libraries_manpages, rather than two
like before.

Co-authored-by: David Allsopp <david.allsopp@metastack.com>