optimize Dynarray.of_list
I missed a natural optimization of Dynarray.of_list, which is to
use List.length to guess the right size for the resulting dynarray
instead of using the dynamic resizing strategy.

A microbenchmark suggests that this is always a win, small (10%
speedup) on small lists and large (2x-3x faster) on large lists.