Fix race between verify_pool and the marking phase (fixes #12902)
`verify_pool` is a function used to check the invariant that a major
pool should contain no garbage after sweeping. It is obviously allowed
to race with marking (which can only turn UNMARKED objects into MARKED
or NOT_MARKABLE, but not into GARBAGE), we only need to mark the read as
a relaxed atomic read.