error messages: reopen input files when given binary AST
We recently simplified the compiler logic to quote source file
fragments, by keeping the whole input file in memory instead of
reopening the input file when the error lies outside the lexing
buffer. This fixes issue with source files that contain preprocessor
directives (typically, they are the output of a source-to-source
preprocessor).

This simplification is a regression for build systems like Dune that
perform a source-to-AST preprocessing step, then call the compiler
with the binary AST as input: in this case there is no source input file in
memory when printing errors, so source quotations are gone.

The present commit restores the fallback logic that we had in 5.1 and
earlier to apply in this case, when given AST as input.