recmod: check consistency before subtyping (#12959)
This commit adds a consistency step just before the last stage of the
typechecking of recursive module bindings:

    module rec M_1: Decl_1 = Impl_1
    and M_2 : Decl_2 = Impl_2
    ...

This new test check that the actual type Actual_i are consistent with
the declared types Decl_i before trying to establish than

   Actual_1, ... , Actual_n <: Decl_1, ..., Decl_n

This step uses a very coarse-grained consistency relation `C`
defined like `<:` except that inclusion for type declarations
and value declarations are replaced by a consistency test, and
all declarations of classes, class types and extensions constructors
are seen as consistent.

For instance,

   sig type -'a t end C sig type 'a t = A end
   sig val x: int end C sig external x: float ="try" end

( `C` is an equivalence on types ), but

  sig type 'a t end ₡ sig type 'a t end

This step ensures that it is safe to add the equation `M_1` = `Impl_1` to
the environment. Without this step, the core system typechecker may fail
with an internal error whenever type declarations are inconsistent
between an implementation and its declared module type.