matching: clarify our static knowledge of the argument-passing logic
The "arguments" in pattern-matching compilation are a list of
scrutinees that are being matched upon, that can most of the time be
formed of arbitrary expressions. The first argument is bound to a
variable when we are about to generate code to switch on the
corresponding column, that is, in the
`compile_match_{nonempty,simplified}` functions. When new arguments
are generated by applying projections to an existing argument, it
is (almost) always on this bound-variable form. In other words, most
arguments can be complex expressions, except the first argument in
some parts of the pattern-matching logic.

The code relies on this simple-argument structure for the first
argument, but would not encode the corresponding invariants in the
type. This is what the current commit does; we can get rid of some
runtime assertions, and simplify code that was more complex than
necessary -- in `compile_match_simplified` in particular.