Support remote debugging over IPv6, and Unix domain sockets on Win32
Support IPv6 on all platforms in ocamldebug CAML_DEBUG_SOCKET.
Passing an IPv6 address (e.g., `[::1]:8080`), or a hostname resolving
to an IPv6 address, is supported.

Support Unix domain sockets on Windows.

Detecting the type of the address is on done on a best-effort basis.
`[:` is a valid DOS drive and counts as an absolute path. We chose to
ignore that and recognize the string as an IPv6 address.