stdlib/set.ml: new submodule [Set.Enum].
This submodule exposes the type [enumeration], which already exists,
as an abstract type of enumerations (sequences) of elements.

Beside the obvious operations that allow iteration ([is_empty],
[head], [tail]), this type supports the operation [from], which
allows efficiently skipping ahead in the sequence to the point
where a certain threshold is reached or exceeded.

This can be used, for instance, to efficiently iterate over the
intersection of two sets without explicitly constructing this
intersection in memory.