sanitizers(MSAN): disable uninstrumented external libraries
Tested with clang-18.1.6 on Fedora 40.

According https://github.com/google/sanitizers/wiki/MemorySanitizer#using-instrumented-libraries
all the code in the program must be built with it, otherwise false
positives are reported.
Disable zstd when compiling with the memory sanitizer, because this is
an external library, that is unlikely to have been compiled with it.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>