sanitizers: use -Og -fno-optimize-sibling-calls
https://clang.llvm.org/docs/MemorySanitizer.html and https://clang.llvm.org/docs/AddressSanitizer.html
both recommend using `-O1 -fno-optimize-sibling-calls` for better
stacktraces, and recommend -O1 or higher for reasonable performance.

Try to future proof this by using -Og, which is currently documented in
`clang` as equivalent to `-O1`, but may in the future disable
optimizations to improve debuggability. GCC supports this flag too, and
infact does disable some -O1 optimizations that interfere with
debugging.

MSAN was using -O0 which is slower than needed.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>