Let configure generate otherlibs/dynlink/dynlink_platform_intf.mli
otherlibs/dynlink/dynlink_platform_intf.mli is a copy of the
corresponding otherlibs/dynlink/dynlink_platform_intf.ml file.
Before this commit, the copy was done during the build. With this commit
the copy (more likely a symlink) is done by configure as is already
done for other files that need to be duplicated.