avoid excessive GC marking during unmarshalling ramp-up phases
When tracking a Coq performance regression in OCaml 5 (#13300), we
realized that the GC work-computation heuristics are bad for ramp-up
phases of programs that do a lot of unmarshalling, such as the loading
of Coq .vo files implied by its `Require Import` directives.

The problem is that the GC assumes a steady state, where it should
work to release approximately the same amount of memory that was
allocated since the beginning of the cycle. When allocating a large
amount of long-lived memory, this assumption results in excessive
marking work (traversing the heap to find free memory).

The fix introduces a sub-category of `major_allocated_words` (the number
of words allocated into the major heap), called
`major_allocated_words_longlived`, which counts the major-heap memory
that we expect to live until the end of time, or the next request for
a full major GC, and is not taken into account to decide marking work.

The only place where this new "longlived" category is used in this
commit is in the unmarshalling code, only in the case where the
post-unmarshalling data is larged than Max_young_wosize (cannot be
allocated as a single block in the minor heap), that is currently 2Mio
on a 64bit system.

This heuristic could be a problem for OCaml programs that allocate a
lot of short-lived memory through unmarshalling, by packets of more
than 2Mio each. Those programs could suffer from vastly increased
memory consuptmion.

Performance numbers for this change, with the Coq default GC settings:

```
Summary
  coqc.5.2+backport+change ran
    1.22 ± 0.02 times faster than coqc.5.2+backport
    1.29 ± 0.10 times faster than coqc.5.2
```

with the OCaml default GC settings:

```
Summary
  coqc.5.2+backport+change ran
    1.30 ± 0.03 times faster than coqc.5.2+backport
    1.41 ± 0.11 times faster than coqc.5.2
```

In these numbers:

- coqc.5.2 is with the stock 5.2 runtime
- coqc.5.2+backport has a backport of #13086,
  which changed the GC pacing slightly
- coqc.5.2+backport+change is the described
  change, on top of the backport