Avoid using C23 `[[]]` syntax for attributes if C23 isn't used
> The special operator `__has_c_attribute (operand)` may be used in
> `#if` and `#elif` expressions in C code to test whether the
> attribute referenced by its operand is recognized by GCC in
> attributes using the `[[]]` syntax.

https://gcc.gnu.org/onlinedocs/cpp/_005f_005fhas_005fc_005fattribute.html

Turns out __has_c_attribute is also defined by GCC and clang even
when they don't default to the C23 standard, even when specifying an
older (e,g,. C17) standard, even with the `-pedantic` flag.

In short, relying on this macro to avoid using the [[]] before C23 is
doomed to fail. Prefer checking whether the compiler is actually using
C23. The attributes are all mandated by the standard.