Fix tests/lib-threads/beat.ml on macOS ARM
Posix specifically notes that nanosleep may delay the calling thread for
_longer_ than the requested timeout while also stating that it may only
return early on error or owing to delivery of a signal.

macOS (especially on Apple Silicon) appears to take a very liberal view
of sub-second nano-sleep calls, such as the the 0.33333 and 0.5 delays
in the beat.ml test.

The test is tweaked to time each Thread.delay (using Unix.gettimeofday)
and allow any extra delay to be credited against the next delay. An
unfair OS scheduler (and a malfunctionaling OCaml runtime) thus will
both still trigger a failure, but a slightly lazy OS scheduler which
inflates some of the nanosleep calls by 30% or so will not cause the
test to fail.