Implement primitive aliases (RFC #44)
This commit implements primitive aliases (RFC #44). Explicitly, we parse
two new forms in both structures and signatures:
- [external p = M.x]
- [external p : t = M.x]

In both cases, the C function names and native representation of
arguments are copied from [M.x], rather than requiring the user to
repeat them in full.

In the first case, the type is copied from [M.x].

In the second case:
- In structures, the supplied type is unified with that of [M.x].
- In signatures, correctness is enforced by the inclusion check.