Add cross-compilation cases to `sak`
When building an OCaml cross-compiler Windows to Unix, `sak` gets a
Windows string (of UTF-16 `wchar_t`s) and must produce a Unix string (of
UTF-8 `char`s). And vice versa when building a cross-compiler Unix to
Windows. To make this possible, this commit splits the
`encode-C-literal` command into two commands with specific encodings of
the result: `encode-C-utf8-literal` and `encode-C-utf16-literal`.

Instead of pulling in a library (and the problems of linking with it)
for the task, this commit adds the specific and simple UTF-* encoders
and decoders that are needed and uses them only when building a
cross-compiler (the result is unchanged in non-cross-compiler settings).

In the cross-compiler UTF-8 case, the non-printable characters are
encoded to get a safer generated code.