Implement the --disable-ocamlobjinfo option for the configure script
This allows to disable the build of the ocamlobjinfo tool.
When used in conjunction with --disable-native-compiler,
it makes it possible to avoid the build of Flambda completely
while one wants to test only the bytecode compiler.

Indeed, when building only the byte code compiler, it's only because of
its use in ocamlobjinfo that Flambda gets built.