use -dmatchcomp to enable matching debug output
Note: -dmatchcomp does not behave like most other -dfoo options,
because it shows course-of-compilation debug output instead of
printing an intermediate representation of a compilation pass -- there
are no separate intermediate representations for pattern-matching
compilation, unfortunately.

In particular, with the current implementation, -dmatchcomp does not
respect the -dump-into-* parameters, it always print to debug
output. Fixing this would be nice, but it requires cumbersome
implementation work, because the matching.ml codebase is not
parametrized over a debug channel.

If someone wanted to tackle this, we would suggest the following route:

- refactor Translcore to turn the ~scopes parameter into a more
generic ~ctx parameter, a record with a `scopes`field
- same thing with Translmod
- add a `ppf_dump` field to this compilation context
- add a `ctx` parameter (at a different type) to all Matching
  functions as well, and pass the `ppf_dump` alongside it
- pass the context to the `debugf` calls