Simplify Parsetree change action and add ppxlib maintainers mention
This slighly simplifies the parsetree change action by making use of
some of the github actions features and the github-script action to
interact with github's API rather than doing it via a separate script.

It also includes a few other files to monitor since ppxlib makes uses
of other parts of compiler-libraries, in addition to `parsetree.mli`,
it now also adds the label if the following files are modified:
- `parsing/location.mli`
- `parsing/longident.mli`
- `parsing/asttypes.mli`
- `parsing/parse.mli`
- `parsing/pprintast.mli`

Finally, since github notifications don't allow tracking a specific
label, the action will ping myself and Patrick whenever a change that
can impact ppxlib happens. That will help us stay in touch with the
latest state of the compiler and work with the contributors to keep
ppxlib's main branch compatible with the development version of OCaml.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>