Simplify output-obj/output-complete-exe machinery
The bytecode linker is now no longer responsible for generating actual
functions - only data is synthesised in bytecomp/bytelink.ml.
Indirectly, this fixes a Windows command-line parsing bug with
-output-complete-exe, as the implementation of main_os in bytelink.ml
was missing a call present in main.c in the libcamlrun.

With this change, libcamlrun now always contains the caml_startup family
of functions, just as libasmrun does. This is intended to simplify the
use of bindings to other language runtimes when loading interop code in
bytecode.