Co-opt prims.c for the bytecode image
The linking trick works beautifully, except that it breaks the MSVC
trick used to workaround around the lack of partial linking. This way,
the symbol is only ever defined once.

This alternate approach treats the bytecode image globals in the same
way as the primitives table - the symbols are not present in libcamlrun
but have to be supplied in the build when linking ocamlrun, etc. (which
is done with prims.o) and similarly when building runtimes with -custom.