Missing configure flag for BasicCompiler on tags
If the Hygiene workflow is triggered for a release tag (or for a branch
sat on a release tag) then ocamltest is not built by default, which
causes the .depend file to change. Explicitly enable ocamltest, even
though it is not built in BasicCompiler, so that the dependency
generation is consistent.