Simplify the checking on caml_global_data
Partly owing to an old multicore API removed prior to merging,
caml_global_data was initialised to NULL in fiber.c and changed to
Val_unit when starting up the interpreter prior to being registered as a
root. That required checks for both NULL and Val_unit to determine if
the global data had been loaded when raising exceptions during
initialisation.

Code updated to be more as the 4.14 version with caml_global_data set to
Val_unit statically.