Treat variables as rigid and use GADT unification
Rather than unifying the type ascribed to a primitive alias with that of
the aliased primitive, we use [Ctype.matches] to check whether the
former is an instance of the latter. Since this might reasonable fail in
signatures, where certain type equations are hidden, we introduce a new
function [Ctype.matches_gadt], which uses the extant GADT unification
logic to check whether the former _could be_ and instance of the latter.

Prior to this commit, the GADT unification logic only ever ran in
structures, and a small change to [Ctype.non_aliasable] is required, to
effectively ignore its result in signatures.